python/vllm/requirements.txt
fastapi
uvicorn
pydantic
google-cloud-aiplatform
vertexai
anyio
asgiref==3.6.0
google-cloud-secret-manager